fix(i/archival): resolve the remaining TODOs
These TODOs are related to existing fields I didn't know how to
fill, so I searched the tree to figure that out.

The largest contributor to this diff is the change the make a
model.HTTPTransport aware of whether it's using "tcp" or "quic".

This change required a bunch of follow-up changes. Many of
these changes were in private FakeTransports that haven't been
converted to using mocks.HTTPTransport already.