chore: upgrade psiphon and go-cmp (#669)
I have experimented with a new approach for embedding psiphon in
https://github.com/ooni/psiphon/commit/7fc0bcd97c5fecd9a8db0c1c01a10c51e5377a40.

It seems the build is still building and the experiment is still
running. With the new approach, we're now vendoring less dependencies,
which hopefully puts us in the right track to, one day, import
Psiphon as a normal Go dependency.

I'll make sure to report to the Psiphon team what is currently
preventing us from importing their ClientLibrary directly.

This work is part of https://github.com/ooni/probe/issues/1894.

As part of running the update, I run `go get -u -v ./...`, which
led to go-cmp also being updated in the process.