cleanup(websteps): remove the concept of oddity
The most recent round of websteps review included feedback saying
that the concept of oddity is confusing to expose in the data.

I think we need some internal-ish form of synthetically saying what
is wrong, but exposing it is perhaps not a good idea.

Let's back out the concept of oddity for now. We will figure out
our needs in a better way when we implement an algorithm to say
what is wrong with a measurement.

Reference issue: https://github.com/ooni/probe/issues/1907