feat(archival): add functionality to wrap netxlite types
My intent here is to basically allow us to reimplement measurex
using functionality in archival, which can then be shared not
only by measurex but also by any other piece of code in the tree.

To this end, I'm basically porting the patches to wrap the
fundamental netxlite types from measurex to archival. The next
step will instead be that of really rewriting measurex to be
based off the code inside archival.