feat(torsf): collect tor logs, select rendezvous method
This diff contains significant improvements over the previous
implementation of the torsf experiment.

We add support for configuring different rendezvous methods after
the convo at https://github.com/ooni/probe/issues/2004. In doing
that, I've tried to use a terminology that is consistent with the
names being actually used by tor developers.

In terms of what to do next, this diff basically instruments
torsf to always rendezvous using domain fronting. Yet, it's also
possible to change the rendezvous method from the command line,
when using miniooni, which allows to experiment a bit more. In the
same vein, by default we use a persistent tor datadir, but it's
also possible to use a temporary datadir using the cmdline.

Here's how a generic invocation of `torsf` looks like:

```bash
./miniooni -ODisablePersistentDatadir=true \
           -ORendezvousMethod=amp \
           -ODisableProgress=true torsf
```

(The default is `DisablePersistentDatadir=false` and
`RendezvousMethod=domain_fronting`.)

With this implementation, we can start measuring whether snowflake
and tor together can boostrap, which seems the most important thing
to focus on at the beginning. Understanding why the bootstrap most
often does not converge with a temporary datadir on Android devices
remains instead an open problem for now. (I'll also update the
relevant issues or create new issues after commit this.)

We also address some methodology improvements that were proposed
in https://github.com/ooni/probe/issues/1686. Namely:

- we record the tor version because we include _some_ tor logs;

- we include the bootstrap percentage because of the logs;

- we set the anomaly key correctly.

What remains to be done is the possibility of including Snowflake
events into the measurement, which is not possible until the new
improvements at common/event in snowflake.git are included into a
tagged version of snowflake itself. (I'll make sure to mention
this aspect to @cohosh in https://github.com/ooni/probe/issues/2004.)

It also remains to be done to measure the amount of bytes sent
and received during the bootstrap, which will also probably
be part of a follow-up diff (or even pull request).

I also expect this diff to fail unit and integration tests, at least
because of reduced coverage. This is fine because I plan to adding
missing tests or fixing them as part of a follow-up diff.

If you're reviewing this diff, I'd recommend focusing on (1) whether
we're collecting good enough data for analysis and (2) whether the
data we collect is safe to collect, or we should collect less to err
more onto the safe side.