fix(torsf): address fundamental bug in original impl
The way I initially wrote torsf in this set of patches was broken
because we could only get logs on success.

What's more, I tried to get logs on failure, which crashed torsf.

Luckily we have unit tests and we noticed.

So, rewrite the code such that we always can access logs.

While there, rework torsf unit tests to ensure they cover all the
cases and they're sorted logically into the file.

While there, start addressing @hellais suggestion that we should
get the version of tor from the control port.

After this diff, we still have some broken or missing tests, and I
will work on that later today or next week.