feat: experiment to measure the download speed
We've received some community requests recently to implement an
experiment that helps to detect throttling.

Doing that properly is quite complex, to be honest.

But I think having an initial instrument to figure out what
happens will help informing a proper experiment for that.

So, this commit creates an initial experiment to start thinking
and reasoning out about this kind of problems.