fix(ptx): downgrade yawning/obfs4
It seems the version of obfs4 we're using introduces a regression as
explained at https://github.com/ooni/probe/issues/1986.

I'm working with @ainghazal to fix this issue in the master branch
of this repository. Our understanding is that the elligator algorithm
changed in obfs4, bringing in significant improvements. But we could
also bisect the issue to such an improvement. So, it may be that some
back and forth is required to settle the issue.

For the purpose of this release branch, thus, I've just downgraded
to a previous version to see if this fixes the issue.

Reference issue: https://github.com/ooni/probe/issues/2016.