fix(android): call getaddrinfo and handle its error
The issue at https://github.com/ooni/probe/issues/2029 is fixed
if we directly call getaddrinfo and correctly map its return code.

While the main reason to propose this diff is to fix the above
mentioned issue, we should note that this diff also paves the way
for https://github.com/ooni/probe/issues/1569.

(Of course, regarding https://github.com/ooni/probe/issues/1569,
we don't have the Rcode when we're calling getaddrinfo, but the
spirit of https://github.com/ooni/probe/issues/1569 is that we
should include the lowest-level error we have seen and, when we're
calling getaddrinfo, such an error is getaddrinfo's retval.)