fix(netxlite): hotfix for Android EAI_NODATA
This cherry-picks 2b48dcfb1a820b2ef28a7d478c3c94f0f8ef4e29 for
the release/3.15 branch. Original commit message follows:

- - -

The problem is explained in https://github.com/ooni/probe/issues/2029.

I am also working on a more comprehensive fix for this
issue in https://github.com/ooni/probe-cli/pull/698.

This diff WILL NOT need forwardporting. It's just meant as
an hotfix for the release/3.14 branch and it's not such that
I'd be happy keeping it in release/3.15+.