refactor(getaddrinfo): propagate runtime.GOOS
We only need to know and use the GOOS inside Android code.

However, to properly test the code, we need toError to receive
the GOOS as an argument, so we can write specific tests for
Android that properly handle EAI_NODATA.

I could possibly only test the specific function below toError
that handles Android but I'd rather test that toError as a whole
is WAI. A future refactoring MAY make Linux code not call such
a specific function. In such a case, with a less precise unit
test, we would never know we've broken the codebase.